hotel-management-service/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
aiofiles